getblocktemplate: Give pre-segwit miners empty templates once segwit has activated
Since throwing an error could cause such a miner to failover to an old node/pool, it seems better to have them use this up-to-date, fully-verifying node even if it means producing empty blocks.
We therefore give them an empty block, and add a bogus inherenty-unsupported "rule" so they know not to try to merge the template.
The "segwit" name is not safe to use in the rules list, in case the template somehow gets to a real segwit-aware miner (eg, a proxy of some sort which may not understand witness transactions).