Remove SigVersion from sighash computation.
Summary:
Trigger BIP143 on bit 6 in nHashType. It's not an issue, because it is illegal as this point in time. This will be extended to ensure replay protection.

The bit is called FORKID. In a future patch, we will add a fork id in the computation of the sighash in order to ensure replay protection.

Test Plan:
  make check
  ../qa/pull-tester/rpc-tests.py

Also updated expectation for sighash tests so that the do not use bit 6.

Reviewers: freetrader, sickpig, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D64