[rpc] distinguish between vsize and mempool_vsize which includes sigop size
CTxMemPoolEntry::GetTxSize() returns the maximum between vsize and
"sigop size" which is used internally.  Users of the RPC interface are
probably not expecting this result from "vsize" (as it is not
documented). Fix this, and add a separate result for "mempool_vsize".

Github-Pull: #27591
Rebased-From: 60bde2dac05bd44b224d6861f6083096c298a2dc